fix(ci): python-coverage-shards — queue PR-reuse GraphQL stubs after tip revalidation
Reuse push paths now fetch live forge snapshots before/after push; FakeCommandRunner
queues still returned git SHAs for those gh api graphql calls, so PR_STATE_LOOKUP_FAILED
failed reuse/recovery tests, and oversized test modules tripped the 1500-line gate.
