+++ tests/unit/control/test_executor_error_paths_parts/test_executor_error_paths_part_025.py
@@ -1,0 +1,757 @@
+"""PR reuse tip-containment and monitor-factory coverage split from part_004."""
+
+from __future__ import annotations
+
+from pathlib import Path
+from types import SimpleNamespace
+from typing import Any
+
+import pytest
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
+
+from awf.common.commands import FakeCommandRunner
+from awf.common.forge_lifecycle import PullRequestLifecycle, PullRequestSnapshot
+from awf.control.executor import helpers as executor_helpers
+from awf.control.executor import pr_open_step as _pr_open_step
+from awf.control.executor.helpers import _call_pr_monitor_factory
+from awf.db.enums import WorkspaceStatus
+from awf.db.repositories import ValidationRunRepository, WorkspaceRepository
+from tests.unit.control.test_executor_error_paths_parts.test_executor_error_paths_part_004 import (
+    _ForgeRecordingPrCreator,
+    _LifecycleForgeClient,
+    _make_executor,
+    _queue_full_happy_path,
+    _queue_pre_push_checks,
+    _queue_validation_head,
+    _seed_ready,
+    factory,
+    fake,
+)
+
+_IMPORTED_FIXTURES = (factory, fake)
+
+
+class TestPullRequestUnexpectedErrorReuseTipPart025:
+    @pytest.mark.unit
+    async def test_reuse_push_fails_closed_when_tip_retry_lookup_fails(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        forge = _LifecycleForgeClient(
+            snapshots=[
+                PullRequestSnapshot(
+                    PullRequestLifecycle.open,
+                    "awf/ws_original",
+                    head_sha="a" * 40,
+                ),
+                PullRequestSnapshot(
+                    PullRequestLifecycle.open,
+                    "awf/ws_original",
+                    head_sha="a" * 40,
+                ),
+                TimeoutError("forge tip-retry timed out"),
+            ]
+        )
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        async def _no_delay(_seconds: float) -> None:
+            return None
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+        monkeypatch.setattr(_pr_open_step, "_POST_PUSH_TIP_RETRY_DELAY_SECONDS", 0.0)
+        monkeypatch.setattr(_pr_open_step.asyncio, "sleep", _no_delay)
+
+        pr_creator = _ForgeRecordingPrCreator()
+        ws_id = await _seed_ready(factory)
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = "https://github.com/x/y/pull/55"
+            ws.pr_number = 55
+            ws.remote_push_branch = "awf/ws_original"
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.snapshot_calls == 3
+        assert len(pr_creator.calls) == 1
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.failed.value
+            assert ws.pr_url == "https://github.com/x/y/pull/55"
+
+    @pytest.mark.unit
+    async def test_reuse_push_keeps_pr_when_post_push_merge_contains_tip(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        # Push then concurrent merge of that same tip: monitor ShortCircuit is OK.
+        pushed_tip = "b" * 40
+        forge = _LifecycleForgeClient(
+            snapshots=[
+                PullRequestSnapshot(
+                    PullRequestLifecycle.open,
+                    "awf/ws_original",
+                    head_sha="a" * 40,
+                ),
+                PullRequestSnapshot(
+                    PullRequestLifecycle.merged,
+                    "awf/ws_original",
+                    head_sha=pushed_tip,
+                ),
+            ]
+        )
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        pr_creator = _ForgeRecordingPrCreator()
+        ws_id = await _seed_ready(factory)
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = "https://github.com/x/y/pull/55"
+            ws.pr_number = 55
+            ws.remote_push_branch = "awf/ws_original"
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.snapshot_calls == 2
+        assert len(pr_creator.calls) == 1
+        assert pr_creator.existing_pr_url == "https://github.com/x/y/pull/55"
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.completed.value
+            assert ws.pr_url == "https://github.com/x/y/pull/55"
+
+    @pytest.mark.unit
+    async def test_reuse_push_opens_replacement_when_post_push_merge_excludes_tip(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        # Snapshot open → push → concurrent merge of the *old* tip: must replace.
+        forge = _LifecycleForgeClient(
+            snapshots=[
+                PullRequestSnapshot(
+                    PullRequestLifecycle.open,
+                    "awf/ws_original",
+                    head_sha="a" * 40,
+                ),
+                PullRequestSnapshot(
+                    PullRequestLifecycle.merged,
+                    "awf/ws_original",
+                    head_sha="a" * 40,
+                ),
+            ],
+            create_url="https://github.com/x/y/pull/99",
+        )
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        pr_creator = _ForgeRecordingPrCreator(
+            new_pr_url="https://github.com/x/y/pull/99",
+        )
+        ws_id = await _seed_ready(factory)
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = "https://github.com/x/y/pull/55"
+            ws.pr_number = 55
+            ws.remote_push_branch = "awf/ws_original"
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.snapshot_calls == 2
+        assert len(pr_creator.calls) == 2
+        assert pr_creator.calls[0]["existing_pr_url"] == "https://github.com/x/y/pull/55"
+        assert pr_creator.calls[1]["existing_pr_url"] is None
+        assert pr_creator.calls[1]["forge_client"] is forge
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.completed.value
+            assert ws.pr_url == "https://github.com/x/y/pull/99"
+            assert ws.pr_number == 99
+
+    @pytest.mark.unit
+    async def test_reuse_push_opens_replacement_when_post_push_pr_closed(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        forge = _LifecycleForgeClient(
+            snapshots=[
+                PullRequestSnapshot(PullRequestLifecycle.open, "awf/ws_original"),
+                PullRequestSnapshot(PullRequestLifecycle.closed, "awf/ws_original"),
+            ],
+            create_url="https://github.com/x/y/pull/99",
+        )
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        pr_creator = _ForgeRecordingPrCreator(
+            new_pr_url="https://github.com/x/y/pull/99",
+        )
+        ws_id = await _seed_ready(factory)
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = "https://github.com/x/y/pull/55"
+            ws.pr_number = 55
+            ws.remote_push_branch = "awf/ws_original"
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.snapshot_calls == 2
+        assert len(pr_creator.calls) == 2
+        assert pr_creator.calls[1]["existing_pr_url"] is None
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.completed.value
+            assert ws.pr_url == "https://github.com/x/y/pull/99"
+
+    @pytest.mark.unit
+    async def test_reuse_push_fails_closed_when_post_push_lookup_fails(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        forge = _LifecycleForgeClient(
+            snapshots=[
+                PullRequestSnapshot(PullRequestLifecycle.open, "awf/ws_original"),
+                TimeoutError("forge lifecycle timed out after push"),
+            ]
+        )
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        pr_creator = _ForgeRecordingPrCreator()
+        ws_id = await _seed_ready(factory)
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = "https://github.com/x/y/pull/55"
+            ws.pr_number = 55
+            ws.remote_push_branch = "awf/ws_original"
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.snapshot_calls == 2
+        assert len(pr_creator.calls) == 1
+        assert pr_creator.existing_pr_url == "https://github.com/x/y/pull/55"
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.failed.value
+            assert ws.pr_url == "https://github.com/x/y/pull/55"
+
+    @pytest.mark.unit
+    async def test_reuse_push_opens_replacement_when_existing_pr_merged(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        # If the preserved PR merged after admission, abandon reuse and open a
+        # replacement so the monitor does not ShortCircuitCompleted on a tip that
+        # never landed in the merged PR.
+        forge = _LifecycleForgeClient(lifecycle=PullRequestLifecycle.merged)
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        pr_creator = _ForgeRecordingPrCreator(
+            new_pr_url="https://github.com/x/y/pull/99",
+        )
+        ws_id = await _seed_ready(factory)
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = "https://github.com/x/y/pull/55"
+            ws.pr_number = 55
+            ws.remote_push_branch = "awf/ws_original"
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.lifecycle_calls == 1
+        assert pr_creator.existing_pr_url is None
+        assert pr_creator.remote_branch_name is None
+        assert pr_creator.forge_client is forge
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.completed.value
+            assert ws.pr_url == "https://github.com/x/y/pull/99"
+            assert ws.pr_number == 99
+            assert ws.remote_push_branch is not None
+
+    @pytest.mark.unit
+    async def test_reuse_push_fails_closed_when_lifecycle_lookup_fails(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        forge = _LifecycleForgeClient(
+            lookup_error=TimeoutError("forge lifecycle timed out"),
+        )
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        pr_creator = _ForgeRecordingPrCreator()
+        ws_id = await _seed_ready(factory)
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = "https://github.com/x/y/pull/55"
+            ws.pr_number = 55
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.lifecycle_calls == 1
+        assert pr_creator.calls == []
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.failed.value
+
+    @pytest.mark.unit
+    async def test_reuse_push_fails_closed_when_pr_number_unresolvable(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        # pr_url without an extractable number must fail closed like a lifecycle
+        # lookup failure: keep identity and do not open a replacement PR.
+        from awf.control.executor.constants import (
+            _AUDIT_PR_CREATED_EVENT,
+        )
+        from awf.control.executor.pr_open_step import (
+            _PR_STATE_LOOKUP_FAILED_REASON_CODE,
+        )
+
+        forge = _LifecycleForgeClient(lifecycle=PullRequestLifecycle.open)
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        pr_creator = _ForgeRecordingPrCreator()
+        ws_id = await _seed_ready(factory)
+        opaque_pr_url = "https://forge.example/x/y/merge-requests/55"
+        async with factory() as s:
+            repo = WorkspaceRepository(s)
+            ws = await repo.get(ws_id)
+            assert ws is not None
+            ws.pr_url = opaque_pr_url
+            ws.pr_number = None
+            ws.remote_push_branch = "awf/ws_original"
+            await s.commit()
+
+        _queue_full_happy_path(fake)
+
+        executor = _make_executor(fake, factory, tmp_path, pr_creator=pr_creator)
+        await executor.execute(ws_id)
+
+        assert forge.lifecycle_calls == 0
+        assert pr_creator.calls == []
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.failed.value
+            assert ws.pr_url == opaque_pr_url
+            assert ws.pr_number is None
+            assert ws.remote_push_branch == "awf/ws_original"
+            assert any(
+                event.event_type == _AUDIT_PR_CREATED_EVENT
+                and event.reason_code == _PR_STATE_LOOKUP_FAILED_REASON_CODE
+                for event in ws.events
+            )
+
+    @pytest.mark.unit
+    async def test_validation_target_sha_update_failure_keeps_open_pr(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+    ) -> None:
+        async def _fail_target_sha_update(
+            *,
+            validation_run_id: str,
+            target_head_sha: str,
+        ) -> None:
+            raise RuntimeError("metadata database temporarily unavailable")
+
+        ws_id = await _seed_ready(factory)
+        fake.queue_result(returncode=0, stdout="adapter ok")
+        fake.queue_result(returncode=0, stdout="awf/x\n")
+        fake.queue_result(returncode=0)
+        fake.queue_result(returncode=0, stdout="")
+        fake.queue_result(returncode=0, stdout="1\n")
+        fake.queue_result(returncode=0)
+        _queue_validation_head(fake)
+        fake.queue_result(returncode=0, stdout="tests ok")
+        _queue_pre_push_checks(fake)
+        fake.queue_result(returncode=0)
+        fake.queue_result(returncode=0, stdout="https://github.com/x/y/pull/7\n")
+
+        executor = _make_executor(fake, factory, tmp_path)
+        executor._set_validation_run_target_head_sha = _fail_target_sha_update  # type: ignore[method-assign]
+
+        await executor.execute(ws_id)
+
+        async with factory() as s:
+            ws = await WorkspaceRepository(s).get(ws_id)
+            assert ws is not None
+            assert ws.status == WorkspaceStatus.completed.value
+            assert ws.pr_url == "https://github.com/x/y/pull/7"
+            assert ws.pr_number == 7
+            assert ws.monitor_last_commit_sha == "deadbeef01"
+
+            runs = await ValidationRunRepository(s).list_for_workspace(ws_id)
+            assert len(runs) == 1
+            assert runs[0].status == "succeeded"
+            assert runs[0].target_head_sha is None
+
+
+class TestPrMonitorFactoryPath:
+    @pytest.mark.unit
+    def test_monitor_factory_supports_one_two_and_three_argument_forms(self) -> None:
+        adapter = object()
+        profile = object()
+        workspace = SimpleNamespace(id="ws_policy", auto_merge=True)
+
+        def _one_arg(adapter: object) -> tuple[str, object]:
+            return ("one", adapter)
+
+        def _two_arg(adapter: object, profile: object) -> tuple[str, object, object]:
+            return ("two", adapter, profile)
+
+        def _three_arg(
+            adapter: object,
+            profile: object,
+            workspace: object,
+        ) -> tuple[str, object, object, object]:
+            return ("three", adapter, profile, workspace)
+
+        assert _call_pr_monitor_factory(
+            _one_arg,
+            adapter=adapter,
+            profile=profile,
+            workspace=workspace,
+        ) == ("one", adapter)
+        assert _call_pr_monitor_factory(
+            _two_arg,
+            adapter=adapter,
+            profile=profile,
+            workspace=workspace,
+        ) == ("two", adapter, profile)
+        assert _call_pr_monitor_factory(
+            _three_arg,
+            adapter=adapter,
+            profile=profile,
+            workspace=workspace,
+        ) == ("three", adapter, profile, workspace)
+
+    @pytest.mark.unit
+    def test_uninspectable_factory_uses_two_argument_fallback(
+        self, monkeypatch: pytest.MonkeyPatch
+    ) -> None:
+        adapter = object()
+        profile = object()
+        workspace = SimpleNamespace(id="ws_policy", auto_merge=True)
+        calls: list[tuple[object, object]] = []
+
+        def _monitor_factory(adapter: object, profile: object) -> object:
+            calls.append((adapter, profile))
+            return "monitor"
+
+        original_signature = executor_helpers.inspect.signature
+
+        def _signature(callable_: object) -> object:
+            if callable_ is _monitor_factory:
+                raise ValueError("signature unavailable")
+            return original_signature(callable_)
+
+        monkeypatch.setattr(executor_helpers.inspect, "signature", _signature)
+
+        assert (
+            _call_pr_monitor_factory(
+                _monitor_factory,
+                adapter=adapter,
+                profile=profile,
+                workspace=workspace,
+            )
+            == "monitor"
+        )
+        assert calls == [(adapter, profile)]
+
+    @pytest.mark.unit
+    def test_adapter_only_factory_preserves_internal_type_error(self) -> None:
+        """Adapter-only factory body TypeErrors should not be masked."""
+        adapter = object()
+        profile = object()
+        workspace = SimpleNamespace(id="ws_policy")
+        factory_error = TypeError("factory body broke")
+        factory_calls: list[object] = []
+
+        def _monitor_factory(adapter: object) -> object:
+            factory_calls.append(adapter)
+            raise factory_error
+
+        with pytest.raises(TypeError, match="factory body broke") as exc_info:
+            _call_pr_monitor_factory(
+                _monitor_factory,
+                adapter=adapter,
+                profile=profile,
+                workspace=workspace,
+            )
+
+        assert exc_info.value is factory_error
+        assert factory_calls == [adapter]
+
+    @pytest.mark.unit
+    def test_three_arg_factory_preserves_internal_type_error(self) -> None:
+        adapter = object()
+        profile = object()
+        workspace = SimpleNamespace(id="ws_policy")
+        factory_error = TypeError("factory body broke after accepting workspace")
+        factory_calls: list[object] = []
+
+        def _monitor_factory(adapter: object, profile: object, workspace: object) -> object:
+            factory_calls.extend([adapter, profile, workspace])
+            raise factory_error
+
+        with pytest.raises(TypeError, match="accepting workspace") as exc_info:
+            _call_pr_monitor_factory(
+                _monitor_factory,
+                adapter=adapter,
+                profile=profile,
+                workspace=workspace,
+            )
+
+        assert exc_info.value is factory_error
+        assert factory_calls == [adapter, profile, workspace]
+
+    @pytest.mark.unit
+    async def test_factory_builds_monitor_once_and_it_runs(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+    ) -> None:
+        """Line 501: when pr_monitor_factory is provided (not a bare
+        monitor), the executor calls it with the created adapter and
+        drives the resulting monitor's ``run()``."""
+        factory_calls: list[Any] = []
+        monitor_calls: list[dict[str, Any]] = []
+
+        class _FakeMonitor:
+            async def run(
+                self, *, workspace_id: str, compose_project: str, compose_file: Path
+            ) -> None:
+                monitor_calls.append(
+                    {"workspace_id": workspace_id, "compose_project": compose_project}
+                )
+                # Don't transition — let the executor's existing code finish.
+
+        def _monitor_factory(adapter: Any) -> _FakeMonitor:
+            factory_calls.append(adapter)
+            return _FakeMonitor()
+
+        ws_id = await _seed_ready(factory)
+        # Drive the full happy path through agent→commit→validate→push→create PR.
+        fake.queue_result(returncode=0, stdout="adapter ok")  # agent
+        fake.queue_result(returncode=0, stdout="awf/x\n")  # current branch
+        fake.queue_result(returncode=0)  # git add
+        fake.queue_result(returncode=0, stdout="a\n")  # cached diff
+        fake.queue_result(returncode=0)  # git commit
+        fake.queue_result(returncode=0, stdout="1\n")  # rev-list count
+        fake.queue_result(returncode=0)  # merge-base --is-ancestor
+        _queue_validation_head(fake)
+        fake.queue_result(returncode=0)  # validation cmd
+        _queue_pre_push_checks(fake, head="deadbeef")
+        fake.queue_result(returncode=0)  # git push
+        fake.queue_result(returncode=0, stdout="https://github.com/x/y/pull/42\n")  # gh pr create
+
+        executor = _make_executor(fake, factory, tmp_path, pr_monitor_factory=_monitor_factory)
+        await executor.execute(ws_id)
+
+        assert len(factory_calls) == 1  # factory called with adapter exactly once
+        assert len(monitor_calls) == 1  # monitor.run fired
+
+    @pytest.mark.unit
+    async def test_existing_pr_recovery_pushes_and_resumes_monitor_without_duplicate_create(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        monitor_calls: list[str] = []
+
+        class _FakeMonitor:
+            async def run(
+                self, *, workspace_id: str, compose_project: str, compose_file: Path
+            ) -> None:
+                del compose_project, compose_file
+                monitor_calls.append(workspace_id)
+
+        def _monitor_factory(adapter: Any) -> _FakeMonitor:
+            del adapter
+            return _FakeMonitor()
+
+        forge = _LifecycleForgeClient(
+            lifecycle=PullRequestLifecycle.open,
+            head_sha="deadbeef",
+        )
+
+        def _make_lifecycle_client(forge_kind: str, runner: object) -> object:
+            return forge
+
+        monkeypatch.setattr(_pr_open_step, "make_forge_client", _make_lifecycle_client)
+
+        ws_id = await _seed_ready(factory)
+        async with factory() as session:
+            workspace = await WorkspaceRepository(session).get(ws_id)
+            assert workspace is not None
+            workspace.pr_url = "https://github.com/x/y/pull/42"
+            workspace.pr_number = 42
+            await session.commit()
+
+        fake.queue_result(returncode=0, stdout="adapter ok")  # agent
+        fake.queue_result(returncode=0, stdout="awf/x\n")  # current branch
+        fake.queue_result(returncode=0)  # git add
+        fake.queue_result(returncode=0, stdout="a\n")  # cached diff
+        fake.queue_result(returncode=0)  # git commit
+        fake.queue_result(returncode=0, stdout="1\n")  # rev-list count
+        fake.queue_result(returncode=0)  # merge-base --is-ancestor
+        _queue_validation_head(fake)
+        fake.queue_result(returncode=0)  # validation cmd
+        _queue_pre_push_checks(fake, head="deadbeef")
+        fake.queue_result(returncode=0)  # git push
+
+        executor = _make_executor(fake, factory, tmp_path, pr_monitor_factory=_monitor_factory)
+        await executor.execute(ws_id)
+
+        assert monitor_calls == [ws_id]
+        assert all(call.args[:3] != ["gh", "pr", "create"] for call in fake.calls)
+        async with factory() as session:
+            workspace = await WorkspaceRepository(session).get(ws_id)
+            assert workspace is not None
+            assert workspace.status == WorkspaceStatus.monitoring_pr.value
+            assert workspace.pr_url == "https://github.com/x/y/pull/42"
+            assert any(event.reason_code == "PR_UPDATED" for event in workspace.events)
+
+    @pytest.mark.unit
+    async def test_executor_passes_workspace_row_to_three_arg_factory(
+        self,
+        fake: FakeCommandRunner,
+        factory: async_sessionmaker[AsyncSession],
+        tmp_path: Path,
+    ) -> None:
+        factory_workspaces: list[Any] = []
+
+        class _FakeMonitor:
+            async def run(
+                self, *, workspace_id: str, compose_project: str, compose_file: Path
+            ) -> None:
+                return None
+
+        def _monitor_factory(adapter: Any, profile: Any, workspace: Any) -> _FakeMonitor:
+            factory_workspaces.append(workspace)
+            return _FakeMonitor()
+
+        ws_id = await _seed_ready(factory, auto_merge=False)
+        fake.queue_result(returncode=0, stdout="adapter ok")  # agent
+        fake.queue_result(returncode=0, stdout="awf/x\n")  # current branch
+        fake.queue_result(returncode=0)  # git add
+        fake.queue_result(returncode=0, stdout="a\n")  # cached diff
+        fake.queue_result(returncode=0)  # git commit
+        fake.queue_result(returncode=0, stdout="1\n")  # rev-list count
+        fake.queue_result(returncode=0)  # merge-base --is-ancestor
+        _queue_validation_head(fake)
+        fake.queue_result(returncode=0)  # validation cmd
+        _queue_pre_push_checks(fake, head="deadbeef")
+        fake.queue_result(returncode=0)  # git push
+        fake.queue_result(returncode=0, stdout="https://github.com/x/y/pull/42\n")
+
+        executor = _make_executor(fake, factory, tmp_path, pr_monitor_factory=_monitor_factory)
+        await executor.execute(ws_id)
+
+        assert len(factory_workspaces) == 1
+        assert factory_workspaces[0].id == ws_id
+        assert factory_workspaces[0].auto_merge is False
